Graph component almost done
